// src/components/roles-manager/types/enums.ts
export enum DataSubject {
  ALL = 'all',
  USER = 'user',
  STUDENT = 'student',
  ADMIN = 'admin',
  SUPER_ADMIN = 'super_admin',
  TEACHER = 'teacher',
  GUARDIAN = 'guardian',
}

export enum Action {
  Manage = 'manage',
  Create = 'create',
  Read = 'read',
  Update = 'update',
  Delete = 'delete',
}